+++ examples/simple_legacy.rs
@@ -1,8 +1,4 @@
-//! This example showcases a list with many widgets.
-//!
-//! Note that we do not implement `Widget` and `PreRender` on the ListItem itself,
-//! but on its reference. This avoids copying the values in each frame and makes
-//! rendering more performant.
+#![allow(deprecated)]
 use crossterm::{
     event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode, KeyEventKind},
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
@@ -13,15 +9,13 @@ use tui_widget_list::{List, ListState, PreRender, PreRenderContext};
 
 type Result<T> = std::result::Result<T, Box<dyn Error>>;
 
-/// Implement `Widget` on the mutable reference to ListItem
-impl Widget for &mut ListItem<'_> {
+impl Widget for ListItem<'_> {
     fn render(self, area: Rect, buf: &mut Buffer) {
         self.clone().get_line().render(area, buf);
     }
 }
 
-/// Implement `PreRender` on the mutable reference to ListItem
-impl PreRender for &mut ListItem<'_> {
+impl PreRender for ListItem<'_> {
     fn pre_render(&mut self, context: &PreRenderContext) -> u16 {
         if context.index % 2 == 0 {
             self.style = Style::default().bg(Color::Rgb(28, 28, 32));
@@ -34,8 +28,6 @@ impl PreRender for &mut ListItem<'_> {
             self.style = Style::default()
                 .bg(Color::Rgb(255, 153, 0))
                 .fg(Color::Rgb(28, 28, 32));
-        } else {
-            self.prefix = None;
         };
 
         1
@@ -67,15 +59,13 @@ fn main() -> Result<()> {
 }
 
 pub fn run_app<B: Backend>(terminal: &mut Terminal<B>, mut app: App) -> io::Result<()> {
-    // Create the widgets only once
-    let mut items: Vec<_> = (0..300000)
+    let items: Vec<_> = (0..30)
         .map(|index| ListItem::new(Line::from(format!("Item {index}"))))
         .collect();
+    let list = List::from(items);
 
     loop {
-        // Then pass them by reference
-        let list = List::from(items.iter_mut().collect::<Vec<_>>());
-        terminal.draw(|f| ui(f, &mut app, list))?;
+        terminal.draw(|f| ui(f, &mut app, list.clone()))?;
 
         if let Event::Key(key) = event::read()? {
             if key.kind == KeyEventKind::Press {
@@ -90,7 +80,7 @@ pub fn run_app<B: Backend>(terminal: &mut Terminal<B>, mut app: App) -> io::Resu
     }
 }
 
-pub fn ui(f: &mut Frame, app: &mut App, list: List<&mut ListItem>) {
+pub fn ui(f: &mut Frame, app: &mut App, list: List<ListItem>) {
     f.render_stateful_widget(list, f.size(), &mut app.state);
 }